inspector_build/requirements.txt
requests
datetime
boto3
botocore
humanfriendly
redis
python-cachetclient
